rename thrift service to avoid class name collisions
